Rename the new factory to `ApiExceptionFactory` and move it out of the `Standard` HTTP Exception namespace.
